Revert "Integer to Long for all ID"
This reverts commit da808168dd128dd945034b3f600d896f764595ab

+++ src/main/java/com/hsbc/roboadvisor/controller/FundSystemController.java
@@ -61,7 +61,7 @@ public class FundSystemController
     @GetMapping("/funds/{fundId}")
     public ResponseEntity<?> getFund(
             @RequestHeader(value = "x-custid") String customerId,
-            @PathVariable Long fundId) {
+            @PathVariable Integer fundId) {
         _logger.info("Getting fund {} for customer id: {}", fundId, customerId);
 
         Fund fund = fundSystemRequestService.getFund(customerId, fundId);
@@ -89,13 +89,13 @@ public class FundSystemController
 
         BigDecimal totalAsset = new BigDecimal(0);
 
-        Map<Long, BigDecimal> fundsMap = new HashMap<>();
+        Map<Integer, BigDecimal> fundsMap = new HashMap<>();
         List<Fund> funds = fundSystemRequestService.getFunds(customerId);
         funds.forEach(fund -> {
             fundsMap.put(fund.getFundId(),fund.getPrice().getAmount());
         });
 
-        Map<Long, Integer> fundIdUnitMap = new HashMap<>();
+        Map<Integer, Integer> fundIdUnitMap = new HashMap<>();
         List<Portfolio> portfolios = fundSystemRequestService.getPortfolios(customerId);
         portfolios.forEach(portfolio -> portfolio.getHoldings().forEach(holding -> {
             if (!fundIdUnitMap.containsKey(holding.getFundId())) {
@@ -106,7 +106,7 @@ public class FundSystemController
             }
         }));
 
-        for (Long key : fundIdUnitMap.keySet()) {
+        for (Integer key : fundIdUnitMap.keySet()) {
             BigDecimal currentValue = fundsMap.get(key);
             Integer units = fundIdUnitMap.get(key);
             totalAsset = totalAsset.add(currentValue.multiply(new BigDecimal(units)));
